[fix] study에서 pickle로 네이밍 변경
