+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
 	id 'io.spring.dependency-management' version '1.1.4'
 }
 
-group = 'peakle_time'
+group = 'pickle_time'
 version = '0.0.1-SNAPSHOT'
 
 java {
