+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'peakle_time'
+rootProject.name = 'pickle_time'
