--- src/main/java/peakle_time/peakle_time/Image/ImageController.java
@@ -1,4 +1,0 @@
-package peakle_time.peakle_time.Image;
-
-public class ImageController {
-}
