--- src/main/java/peakle_time/peakle_time/Image/ImageRepository.java
@@ -1,8 +1,0 @@
-package peakle_time.peakle_time.Image;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ImageRepository extends JpaRepository<Image,Long> {
-}
