--- src/main/java/peakle_time/peakle_time/Member/MemberRepository.java
@@ -1,14 +1,0 @@
-package peakle_time.peakle_time.Member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    boolean existsByEmail(String email);
-    boolean existsByNickname(String nickname);
-    Optional<Member> findByEmail(String email);
-
-}
