--- src/main/java/peakle_time/peakle_time/Member/dto/JoinRequest.java
@@ -1,12 +1,0 @@
-package peakle_time.peakle_time.Member.dto;
-
-public record JoinRequest(
-
-        String email,
-        String password,
-        String checkPassword,
-        String nickname,
-        String company
-
-) {
-}
