--- src/main/java/peakle_time/peakle_time/Member/dto/UpdateRequest.java
@@ -1,8 +1,0 @@
-package peakle_time.peakle_time.Member.dto;
-
-public record UpdateRequest(
-        String nickname,
-        String email,
-        String company
-) {
-}
