--- src/main/java/peakle_time/peakle_time/PeakleTimeApplication.java
@@ -1,13 +1,0 @@
-package peakle_time.peakle_time;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PeakleTimeApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PeakleTimeApplication.class, args);
-	}
-
-}
