--- src/main/java/peakle_time/peakle_time/Review/Review.java
@@ -1,23 +1,0 @@
-package peakle_time.peakle_time.Review;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import peakle_time.peakle_time.global.BaseEntity;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Review extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "record")
-    private String record;
-
-    @Column(name = "star")
-    private int star;
-}
