--- src/main/java/peakle_time/peakle_time/Review/ReviewRepository.java
@@ -1,8 +1,0 @@
-package peakle_time.peakle_time.Review;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ReviewRepository extends JpaRepository<Review,Long> {
-}
