--- src/main/java/peakle_time/peakle_time/Review/ReviewService.java
@@ -1,4 +1,0 @@
-package peakle_time.peakle_time.Review;
-
-public class ReviewService {
-}
