--- src/main/java/peakle_time/peakle_time/Study/Status.java
@@ -1,5 +1,0 @@
-package peakle_time.peakle_time.Study;
-
-public enum Status {
-    Start,End
-}
