--- src/main/java/peakle_time/peakle_time/Study/StudyController.java
@@ -1,7 +1,0 @@
-package peakle_time.peakle_time.Study;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class StudyController {
-}
