--- src/main/java/peakle_time/peakle_time/Study/StudyRepository.java
@@ -1,8 +1,0 @@
-package peakle_time.peakle_time.Study;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface StudyRepository extends JpaRepository<Study,Long> {
-}
