--- src/main/java/peakle_time/peakle_time/Study/StudyService.java
@@ -1,4 +1,0 @@
-package peakle_time.peakle_time.Study;
-
-public class StudyService {
-}
