--- src/main/java/peakle_time/peakle_time/global/BaseEntity.java
@@ -1,29 +1,0 @@
-package peakle_time.peakle_time.global;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.MappedSuperclass;
-import jakarta.persistence.PreUpdate;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-public abstract class BaseEntity {
-
-    @Column(updatable = false, nullable = false)
-    private LocalDateTime createdAt;
-
-    @Column(nullable = false)
-    private LocalDateTime updatedAt;
-
-    public BaseEntity() {
-        this.createdAt = LocalDateTime.now();
-        this.updatedAt = LocalDateTime.now();
-    }
-
-    @PreUpdate
-    public void update() {
-        this.updatedAt = LocalDateTime.now();
-    }
-}
