+++ src/main/java/pickle_time/pickle_time/Image/Image.java
@@ -1,0 +1,19 @@
+package pickle_time.pickle_time.Image;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Image {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String filename;
+    private String filepath;
+
+}
