+++ src/main/java/pickle_time/pickle_time/Image/ImageController.java
@@ -1,0 +1,4 @@
+package pickle_time.pickle_time.Image;
+
+public class ImageController {
+}
