+++ src/main/java/pickle_time/pickle_time/Image/ImageRepository.java
@@ -1,0 +1,8 @@
+package pickle_time.pickle_time.Image;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ImageRepository extends JpaRepository<Image,Long> {
+}
