+++ src/main/java/pickle_time/pickle_time/Member/MemberController.java
@@ -1,0 +1,70 @@
+package pickle_time.pickle_time.Member;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import pickle_time.pickle_time.Member.dto.JoinRequest;
+import pickle_time.pickle_time.Member.dto.UpdateRequest;
+
+import java.util.List;
+import java.util.Optional;
+
+
+@RestController
+@RequiredArgsConstructor
+public class MemberController {
+
+    private final MemberService memberService;
+
+    @PostMapping("/join")
+    public ResponseEntity<?> join(@RequestBody JoinRequest joinRequest) {
+        try {
+            memberService.join(joinRequest);
+            return ResponseEntity.ok("회원가입 성공");
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<?> login(@RequestParam String email, @RequestParam String password) {
+        try {
+            memberService.login(email, password);
+            return ResponseEntity.ok("로그인 성공");
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<?> getMemberById(@PathVariable Long id) {
+        Optional<Member> member = memberService.findById(id);
+        return member.isPresent() ? ResponseEntity.ok(member.get()) : ResponseEntity.notFound().build();
+    }
+
+    @GetMapping("/all")
+    public ResponseEntity<List<Member>> getAllMembers() {
+        List<Member> members = memberService.findAll();
+        return ResponseEntity.ok(members);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<?> updateMember(@PathVariable Long id, @RequestBody UpdateRequest updateRequest) {
+        try {
+            Member updatedMember = memberService.updateMember(id, updateRequest);
+            return ResponseEntity.ok(updatedMember);
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<?> deleteMember(@PathVariable Long id) {
+        try {
+            memberService.deleteMember(id);
+            return ResponseEntity.ok("회원 삭제 성공");
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
+    }
+}
