+++ src/main/java/pickle_time/pickle_time/Member/MemberRepository.java
@@ -1,0 +1,14 @@
+package pickle_time.pickle_time.Member;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    boolean existsByEmail(String email);
+    boolean existsByNickname(String nickname);
+    Optional<Member> findByEmail(String email);
+
+}
