+++ src/main/java/pickle_time/pickle_time/Member/dto/JoinRequest.java
@@ -1,0 +1,12 @@
+package pickle_time.pickle_time.Member.dto;
+
+public record JoinRequest(
+
+        String email,
+        String password,
+        String checkPassword,
+        String nickname,
+        String company
+
+) {
+}
