+++ src/main/java/pickle_time/pickle_time/Member/dto/UpdateRequest.java
@@ -1,0 +1,8 @@
+package pickle_time.pickle_time.Member.dto;
+
+public record UpdateRequest(
+        String nickname,
+        String email,
+        String company
+) {
+}
