+++ src/main/java/pickle_time/pickle_time/Pickle/Status.java
@@ -1,0 +1,5 @@
+package pickle_time.pickle_time.Pickle;
+
+public enum Status {
+    Start,End
+}
