+++ src/main/java/pickle_time/pickle_time/PickleTimeApplication.java
@@ -1,0 +1,13 @@
+package pickle_time.pickle_time;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PickleTimeApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PickleTimeApplication.class, args);
+	}
+
+}
