+++ src/main/java/pickle_time/pickle_time/Review/Review.java
@@ -1,0 +1,23 @@
+package pickle_time.pickle_time.Review;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import pickle_time.pickle_time.global.BaseEntity;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Review extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "record")
+    private String record;
+
+    @Column(name = "star")
+    private int star;
+}
