+++ src/main/java/pickle_time/pickle_time/Review/ReviewRepository.java
@@ -1,0 +1,8 @@
+package pickle_time.pickle_time.Review;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ReviewRepository extends JpaRepository<Review,Long> {
+}
