--- src/test/java/peakle_time/peakle_time/Member/MemberServiceTest.java
@@ -1,15 +1,0 @@
-package peakle_time.peakle_time.Member;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-@SpringBootTest
-class MemberServiceTest {
-
-
-
-
-
-}
