--- src/test/java/peakle_time/peakle_time/PeakleTimeApplicationTests.java
@@ -1,13 +1,0 @@
-package peakle_time.peakle_time;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PeakleTimeApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
