+++ src/test/java/pickle_time/pickle_time/Member/MemberServiceTest.java
@@ -1,0 +1,15 @@
+package pickle_time.pickle_time.Member;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@SpringBootTest
+class MemberServiceTest {
+
+
+
+
+
+}
