+++ src/test/java/pickle_time/pickle_time/PickleTimeApplicationTests.java
@@ -1,0 +1,13 @@
+package pickle_time.pickle_time;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PickleTimeApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
